It works on heroku if i comment resize line in readImage function...

+++ src/main/java/com/lvlz/gallery/data/learn/ClassifierNative.java
@@ -83,7 +83,7 @@ public class ClassifierNative {
 
   private Mat readImage(String imagePath) {
 
-    double newWidth, newHeight;
+    int newWidth, newHeight;
 
     int shapeChange;
     byte[] imBytes;
@@ -140,10 +140,10 @@ public class ClassifierNative {
 
       }
 
-      newWidth = (double) (image.size(2) * shapeChange);
-      newHeight = (double) (image.size(1) * shapeChange);
+      newWidth = image.size(2) * shapeChange;
+      newHeight = image.size(1) * shapeChange;
 
-      resize(image, image, new Size(newWidth, newHeight), newWidth / ((double) image.size(2)), newHeight / ((double) image.size(1)));
+      resize(image, image, new Size(newWidth, newHeight), (double) (newWidth / image.size(2)), (double) (newHeight / image.size(1)));
 
     }
